cli: move yarn utils into versioning lib
