+++ packages/cli/src/lib/versioning/packages.ts
@@ -17,7 +17,7 @@
 import { minimatch } from 'minimatch';
 import { getPackages } from '@manypkg/get-packages';
 import { NotFoundError } from '../errors';
-import { detectYarnVersion } from '../yarn';
+import { detectYarnVersion } from './yarn';
 import { execFile } from '../run';
 
 const DEP_TYPES = [
